Refactor `cmd` package layout and update import paths
- Move command-related files from `src/cmd` to the root-level `cmd` directory.
- Update `main.go` import path from `github.com/Pengxn/go-xn/src/cmd` to `github.com/Pengxn/go-xn/cmd`.
- Keep behavior unchanged while aligning package structure with a cleaner project layout.
